Removed the players._language module.
